+++ notes/spinta/release/pre_release.sh
@@ -7,9 +7,9 @@ test -n "$PID" && kill "$PID"
 
 # Setup versions and create prepare branch
 export MAJOR=0
-export MINOR=2dev4
-export OLD_MINOR=2dev3
-export FUTURE_MINOR=2dev5
+export MINOR=2dev6
+export OLD_MINOR=2dev5
+export FUTURE_MINOR=2dev7
 export RELEASE_VERSION=$MAJOR.$MINOR
 export CURRENT_VERSION=$MAJOR.$OLD_MINOR
 export FUTURE_VERSION=$MAJOR.$FUTURE_MINOR
@@ -75,7 +75,7 @@ poetry run pytest -vvx --tb=short tests
 #| Did not run this time
 
 # Check if new Spinta version works with manifest files
-poetry shell
+poetry env activate
 
 # Configure Spinta server instance
 INSTANCE=releases/$NEW_VERSION