+++ CHANGES.rst
@@ -21,6 +21,12 @@ Improvements:
   .. _#1462: https://github.com/atviriduomenys/spinta/issues/1462
   .. _#1486: https://github.com/atviriduomenys/spinta/issues/1486
 
+Bug fixes:
+
+- Fixed `spinta copy` ignores resources without any models (`#1512`_)
+
+  .. _#1512: https://github.com/atviriduomenys/spinta/issues/1512
+
 Other:
 - Removed dependency `mypy`