+++ CHANGES.rst
@@ -21,6 +21,13 @@ Improvements:
   .. _#1462: https://github.com/atviriduomenys/spinta/issues/1462
   .. _#1486: https://github.com/atviriduomenys/spinta/issues/1486
 
+Bug fixes:
+
+- Removed `_base` column from HTML response when viewing SOAP data with URL parameters (`#1338`_)
+- Added required parameters validation, when building SOAP query, and raising exception `MissingRequiredProperty` if parameter is missing (`#1338`_)
+
+  .. _#1338: https://github.com/atviriduomenys/spinta/issues/1338
+
 Other:
 - Removed dependency `mypy`