+++ CHANGES.rst
@@ -24,6 +24,13 @@ Improvements:
 Other:
 - Removed dependency `mypy`
 
+New Features:
+
+- Added A template-based OpenAPI generator that creates REST API documentation from manifest. (`#1463`)
+
+  .. _#1463: https://github.com/atviriduomenys/spinta/issues/1463
+
+
 0.2dev5 (2025-09-03)
 ====================