+++ CHANGES.rst
@@ -28,8 +28,6 @@ Improvements:
   .. _#1461: https://github.com/atviriduomenys/spinta/issues/1461
   .. _#1462: https://github.com/atviriduomenys/spinta/issues/1462
   .. _#1486: https://github.com/atviriduomenys/spinta/issues/1486
-  .. _#1506: https://github.com/atviriduomenys/spinta/issues/1506
-  .. _#1338: https://github.com/atviriduomenys/spinta/issues/1338
 
 Bug fixes:
 
@@ -38,6 +36,7 @@ Bug fixes:
 - Added required parameters validation, when building SOAP query, and raising exception `MissingRequiredProperty` if parameter is missing (`#1338`_)
 
   .. _#1512: https://github.com/atviriduomenys/spinta/issues/1512
+  .. _#1338: https://github.com/atviriduomenys/spinta/issues/1338
 
 Other:
 - Removed dependency `mypy`